Move static method to new PayloadUtilities class.
